requirements.txt
gradio>=4.19.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
langchain>=0.1.9
langchain-openai>=0.0.8
langchain-google-genai>=0.0.9
langchain-anthropic>=0.1.4
anthropic>=0.18.1
openai>=1.12.0
google-generativeai>=0.3.2 